mediacapture-streams/MediaStreamTrack-MediaElement-disabled-audio-is-silence.https.html

<!doctype html>
<html>
<head>
<title>A disabled audio track is rendered as silence</title>
<link rel="author" title="Dominique Hazael-Massieux" href="mailto:dom@w3.org"/>
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#introduction">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#mediastreams-as-media-elements">
</head>
<body>
<p class="instructions">When prompted, accept to share your audio stream.</p>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks that a disabled audio track in a
MediaStream is rendered as silence. It relies on the
<a href="https://dvcs.w3.org/hg/audio/raw-file/tip/webaudio/specification.html">
Web Audio API</a>.</p>

<div id='log'></div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
promise_test(t => {
  let aud = document.getElementById("aud");
  return navigator.mediaDevices.getUserMedia({audio: true})
  .then(stream => new Promise(resolve => {
    let ctx = new AudioContext();
    let streamSource = ctx.createMediaStreamSource(stream);
    let silenceDetector = ctx.createScriptProcessor(1024);
    let count = 10;
    silenceDetector.onaudioprocess = e => {
      let buffer1 = e.inputBuffer.getChannelData(0);
      let buffer2 = e.inputBuffer.getChannelData(1);
      let out = e.outputBuffer.getChannelData(0);
      out = new Float32Array(buffer1);
      for (let i = 0; i < buffer1.length; i++) {
        assert_equals(buffer1[i], 0, "Audio buffer entry #" + i + " in channel 0 is silent");
      }
      for (let i = 0; i < buffer2.length; i++) {
        assert_equals(buffer2[i], 0, "Audio buffer entry #" + i + " in channel 1 is silent");
      }
      count--;
      if (count === 0) {
        silenceDetector.onaudioprocess = null;
        resolve();
      }
    };
    stream.getAudioTracks()[0].enabled = false;

    streamSource.connect(silenceDetector);
    silenceDetector.connect(ctx.destination);
  }))
}, "Tests that a disabled audio track in a MediaStream is rendered as silence", {timeout: 200000});
</script>
</body>
</html>
